feat(tailwind): export tailwind preset. Refs #11
